переименовала дом-элументы в main, перенесла необходимые функции из mock
