+++ src/const.js
@@ -1,0 +1,6 @@
+const FilterType = {
+  EVERYTHING: 'everything',
+  FUTURE: 'future',
+};
+
+export {FilterType};
