+++ src/main.js
@@ -2,20 +2,20 @@ import {render} from './framework/render.js';
 import FilterView from './view/filter-view.js';
 import ListPresenter from './presenter/list-presenter.js';
 import PointsModel from './model/points-model.js';
-import {generateFilter} from './mock/point.js';
+import {generateFilter} from './utils.js';
 
-const siteHeader = document.querySelector('.page-header');
-const filters = siteHeader.querySelector('.trip-controls__filters');
-const siteMain = document.querySelector('.page-main');
-const eventsContainer = siteMain.querySelector('.trip-events');
+const siteHeaderEl = document.querySelector('.page-header');
+const filtersEl = siteHeaderEl.querySelector('.trip-controls__filters');
+const siteMainEl = document.querySelector('.page-main');
+const eventsContainerEl = siteMainEl.querySelector('.trip-events');
 
 const pointsModel = new PointsModel();
-const listPresenter = new ListPresenter({listContainer: eventsContainer, pointsModel});
+const listPresenter = new ListPresenter({listContainer: eventsContainerEl, pointsModel});
 
 const filter = generateFilter(pointsModel.points);
 
 
-render(new FilterView(filter), filters);
+render(new FilterView(filter), filtersEl);
 listPresenter.init();
