+++ src/mock/point.js
@@ -1,6 +1,5 @@
-import {CITIES, DESCRIPTION, OFFER_TITLE, PHOTO, POINT_TYPE, FilterType} from './const.js';
+import {CITIES, DESCRIPTION, OFFER_TITLE, PHOTO, POINT_TYPE} from './const.js';
 import {generateId, getRandomArrayElement, getRandomNumber} from '../utils.js';
-import dayjs from 'dayjs';
 
 const POINT_COUNT = 4;
 
@@ -105,20 +104,4 @@ function getRandomMockPoints() {
   return Array.from({length: POINT_COUNT}, getRandomMockPoint);
 }
 
-const isFutureTask = (dateFrom, dateTo) => dayjs().isBefore(dayjs(dateFrom)) || dayjs().isSame(dayjs(dateFrom), 'day') || (dayjs().isAfter(dayjs(dateFrom)) && dayjs().isBefore(dayjs(dateTo)));
-
-const filters = {
-  [FilterType.EVERYTHING]: (points) => points,
-  [FilterType.FUTURE]: (points) => points.filter((point) => isFutureTask(point.dateFrom, point.dateTo)),
-};
-
-function generateFilter(points) {
-  return Object.entries(filters).map(
-    ([filterName, filteredPoints]) => ({
-      name: filterName,
-      count: filteredPoints(points).length,
-    }),
-  );
-}
-
-export {getRandomMockPoints, offersByType, getOffersArray, generateFilter};
+export {getRandomMockPoints, offersByType, getOffersArray};
